INS-320: icc renamed to insgocc
