+++ .gitignore
@@ -38,4 +38,6 @@ vm/wasm/test/data/*.wat
 
 logicrunner/goplugin/ginsider-cli/ginsider-cli
 logicrunner/goplugin/preprocessor/preprocessor
+cmd/icc/icc
+cmd/insgocc/insgocc
 *_generated.go
